#89: make demo version of staypoint engine executable
