+++ dev/tsconfig.json
@@ -15,6 +15,7 @@
       "src/*": ["../src/*"]
     },
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "experimentalDecorators": true
   }
 }
